# challenge5/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "DigiBank Documentation - OpenAPI 3.0",
    "description": "Dokumentasi API DigiBank",
    "contact": {
      "email": "apiteam@swagger.io"
    },
    "version": "1.0.11"
  },
  "servers": [
    {
      "url": "http://localhost:5000/api/v1"
    }
  ],
...
  "paths": {
    "/users": {
      "get": {
        "tags": [
          "users"
        ],
        "summary": "Get users lists",
        "description": "Get users Lists",
        "operationId": "getUsers",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "Pagination page",
            "required": false,
            "explode": true,
            "schema": {
              "type": "number",
              "default": 1
            }
          },
          {
            "name": "itemsPerPage",
            "in": "query",
            "description": "Pagination limit",
            "required": false,
            "explode": true,
            "schema": {
              "type": "number",
              "default": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Users"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "authorization": []
          }
        ]
      },
      "post": {
        "tags": [
          "users"
        ],
        "summary": "Add a new users",
        "description": "Add a new users",
        "requestBody": {
          "$ref": "#/components/requestBodies/Register"
        },
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Register"
                }
              }
            }
          }
        },
        "security": [
          {
            "authorization": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "Users": {
        "type": "object",
        "properties": {
          "status": {
            "type": "integer",
            "example": 200
          },
          "message": {
            "type": "string",
            "example": "SUCCESS"
          },
          "data": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "customer_id": {
                  "type": "integer",
                  "example": 66
                },
                "full_name": {
                  "type": "string",
                  "example": "agus"
                },
                "gender": {
                  "type": "string",
                  "example": "Male"
                },
                "address": {
                  "type": "string",
                  "example": "test"
                },
                "phone_number": {
                  "type": "string",
                  "example": "1231232132"
                },
                "email": {
                  "type": "string",
                  "format": "email",
                  "example": "test@mail.com"
                },
                "date_of_birth": {
                  "type": "string",
                  "format": "date-time",
                  "example": "1990-02-15T00:00:00.000Z"
                },
                "identification_number": {
                  "type": "string",
                  "example": "12312321"
                },
                "created_at": {
                  "type": "string",
                  "format": "date-time",
                  "example": "2023-11-11T14:59:30.848Z"
                },
                "updated_at": {
                  "type": "string",
                  "format": "date-time",
                  "example": "2023-11-11T14:59:30.848Z"
                },
                "is_delete": {
                  "type": "boolean",
                  "example": false
                },
                "username": {
                  "type": "string",
                  "example": "username8977.007392914222"
                }
              }
            }
          }
        }
      },
...
      "Register": {
        "type": "object",
        "properties": {
          "status": {
            "type": "integer",
            "example": 201
          },
          "message": {
            "type": "string",
            "example": "CREATED"
          },
          "data": {
            "type": "object",
            "properties": {
              "customer_id": {
                "type": "integer",
                "example": 749
              },
              "full_name": {
                "type": "string",
                "example": "agus"
              },
              "gender": {
                "type": "string",
                "example": "Male"
              },
              "address": {
                "type": "string",
                "example": "test"
              },
              "phone_number": {
                "type": "string",
                "example": "1231232132"
              },
              "email": {
                "type": "string",
                "format": "email",
                "example": "test@mail.com"
              },
              "date_of_birth": {
                "type": "string",
                "format": "date-time",
                "example": "1990-02-15T00:00:00.000Z"
              },
              "identification_number": {
                "type": "string",
                "example": "12312321"
              },
              "created_at": {
                "type": "string",
                "format": "date-time",
                "example": "2023-11-13T20:12:25.718Z"
              },
              "updated_at": {
                "type": "string",
                "format": "date-time",
                "example": "2023-11-13T20:12:25.718Z"
              },
              "is_delete": {
                "type": "boolean",
                "example": false
              },
              "username": {
                "type": "string",
                "example": "usernameagus"
              },
              "password": {
                "type": "string",
                "example": "$2b$10$d8.gY9cr0PN8kOmwAjqg2OZ9RwLeHkTIpFwtuJ.lRkM3RlgtmVIA2"
              }
            }
          }
        }
      },
...
    },
    "requestBodies": {
...
      "Register":{
        "description": "register body",
        "content":{
          "application/json":{
            "schema":{
              "type": "object",
              "properties": {
                "full_name": {
                  "type": "string",
                  "example": "agus"
                },
                "gender": {
                  "type": "string",
                  "example": "Male"
                },
                "address": {
                  "type": "string",
                  "example": "test"
                },
                "email": {
                  "type": "string",
                  "format": "email",
                  "example": "test@mail.com"
                },
                "date_of_birth": {
                  "type": "string",
                  "format": "date-time",
                  "example": "1990-02-15T00:00:00.000Z"
                },
                "identification_number": {
                  "type": "string",
                  "example": "12312321"
                },
                "username": {
                  "type": "string",
                  "example": "usernameagus"
                },
                "password": {
                  "type": "string",
                  "example": "password"
                },
                "phone_number": {
                  "type": "string",
                  "example": "1231232132"
                }
              },
              "required": ["full_name", "gender", "address", "email", "date_of_birth", "identification_number", "username", "password", "phone_number"]
            }
          }
        }
      },
...
    },
    "securitySchemes": {
      "authorization": {
        "type": "apiKey",
        "name": "authorization",
        "in": "header"
      }
    }
  }
}